+++ BitmovinAnalyticsCollector/Classes/AVPlayer/AVPlayerAdapter.swift
@@ -16,6 +16,7 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     private var currentVideoBitrate: Double = 0
     private var previousTime: CMTime?
     private var isPlayerReady = false
+    internal var currentSourceMetadata: SourceMetadata?
     
     internal var drmDownloadTime: Int64?
     private var drmType: String?