+++ BitmovinAnalyticsCollector/Classes/Collector/config/CustomDataConfig.swift
@@ -1,0 +1,4 @@
+internal protocol CustomDataConfig {
+    func getCustomData() -> CustomData
+    func setCustomData(customData: CustomData)
+}