+++ BitmovinAnalyticsCollector/Classes/Collector/data/CustomData.swift
@@ -1,0 +1,21 @@
+@objc
+@objcMembers
+public class CustomData: NSObject {
+   
+    @objc public var customData1: String?
+
+    @objc public var customData2: String?
+
+    @objc public var customData3: String?
+
+    @objc public var customData4: String?
+
+    @objc public var customData5: String?
+
+    @objc public var customData6: String?
+    
+    @objc public var customData7: String?
+    
+    @objc public var experimentName: String?
+    
+}