+++ BitmovinAnalyticsCollector/Classes/Collector/stateMachines/StateMachine.swift
@@ -275,4 +275,16 @@ public class StateMachine {
     public func setErrorData(error: ErrorData?) {
         self.errorData = error
     }
+    
+    internal func changeCustomData(customData: CustomData, time: CMTime?, customDataConfig: CustomDataConfig){
+        let originalState = state
+        let shouldTransition = (originalState == PlayerState.paused || originalState == PlayerState.playing)
+        if shouldTransition {
+            transitionState(destinationState: PlayerState.customdatachange,time: time)
+        }
+        customDataConfig.setCustomData(customData: customData)
+        if shouldTransition {
+            transitionState(destinationState: originalState, time: time)
+        }
+    }
 }